chore: 미디어 request, response 파일 분리
