--- android/core/network/src/main/kotlin/com/andlife/network/api/media/CompleteUploadRequest.kt
@@ -1,7 +1,0 @@
-package com.andlife.network.api.media
-
-// @Serializable
-// data class CompleteUploadRequest(
-//    val mediaKey: String,
-//    val mediaType: String
-// )
