--- android/core/network/src/main/kotlin/com/andlife/network/api/media/CompleteUploadResponse.kt
@@ -1,8 +1,0 @@
-package com.andlife.network.api.media
-
-// @Serializable
-// data class CompleteUploadResponse(
-//    val success: Boolean,
-//    val mediaUrl: String? = null,
-//    val message: String? = null
-// )
