--- android/core/network/src/main/kotlin/com/andlife/network/api/media/UploadMediaRequest.kt
@@ -1,7 +1,0 @@
-package com.andlife.network.api.media
-
-// @Serializable
-// data class UploadMediaRequest(
-//    val mediaType: String,  // "VIDEO", "IMAGE", "AUDIO"
-//    val fileName: String
-// )
